Added ARTParser - Parsing ART functions as a scene Color now stored as a double to match other types Added Validate button which parses the ART file on the left hand side Running the render also parses the active ART file Errors of the ARTParser get output to the bottom left of the window Updated README to document the ART functions Added licenses to Credits of the README

+++ AnotherRayTracer/Color.cpp
@@ -1,6 +1,6 @@
 #include "Color.h"
 
-Color::Color(float _r, float _g, float _b) : r(_r), g(_g), b(_b) {}
+Color::Color(double _r, double _g, double _b) : r(_r), g(_g), b(_b) {}
 
 Color Color::operator+(Color other) const
 {
